backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
sqlalchemy==2.0.28
pydantic==2.6.4
python-dotenv==1.0.1
supabase==2.4.0
motor==3.3.2
python-jose[cryptography]==3.3.0
twilio==9.3.0
langchain==0.1.12
openai==1.12.0
pandas==2.2.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
reportlab==4.1.0
pytest==8.0.2
httpx==0.27.0
